+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# Tamr Azure Databricks module - v0.2.0 - May 12th 2020
+* Updates the module to be compliant with Azure provider 2.11.0
+* Adds a minimal example
+
 # Tamr Azure Databricks module - v0.1.0 - May 12th 2020
 ## Added
 * Tamr Azure Databricks module added